+++ R/create_base_env.R
@@ -2,7 +2,7 @@
 #' @keywords internal
 #' @noRd
 create_base_env <- function(verbose = "silent") {
-  if (isFALSE(env_exists(env_name = "condathis-env"))) {
+  if (isFALSE(env_exists(env_name = "condathis-env", verbose = verbose))) {
     create_env(
       env_name = "condathis-env",
       verbose = verbose,