+++ R/parse_quiet_flag.R
@@ -1,5 +1,7 @@
 #' Parse Quiet Flag From Verbosity Level
 #'
+#' DEPRECATED: As of version 0.1.3, this function is deprecated. Please use the updated verbosity handling functions.
+#'
 #' @param verbose Character string specifying the verbosity level.
 #' @keywords internal
 #' @noRd