+++ tests/testthat/test-rethrow_error.R
@@ -10,7 +10,11 @@ testthat::test_that("Do not execute code in curly braces", {
   testthat::expect_error(
     object = {
       run(
-        "R", "-q", "-s", "-e", "stop(\"{ 5 + x }\")",
+        "R",
+        "-q",
+        "-s",
+        "-e",
+        "stop(\"{ 5 + x }\")",
         env_name = "condathis-test-env",
         verbose = "silent",
         error = "cancel"
@@ -22,7 +26,11 @@ testthat::test_that("Do not execute code in curly braces", {
   testthat::expect_error(
     object = {
       run(
-        "R", "-q", "-s", "-e", "stop(\"{{{ 5 + x }}}\")",
+        "R",
+        "-q",
+        "-s",
+        "-e",
+        "stop(\"{{{ 5 + x }}}\")",
         env_name = "condathis-test-env",
         verbose = "silent",
         error = "cancel"
@@ -34,7 +42,11 @@ testthat::test_that("Do not execute code in curly braces", {
   px_res <- rethrow_error_run(
     expr = {
       run(
-        "R", "-q", "-s", "-e", "stop(\"{{{ 5 + x }}}\")",
+        "R",
+        "-q",
+        "-s",
+        "-e",
+        "stop(\"{{{ 5 + x }}}\")",
         env_name = "condathis-test-env",
         verbose = "silent",
         error = "continue"